Update Mother Brain, Ridley và fix lại cho đạn và nhiều thứ khác

+++ Metroid/Bedgehog.cpp
@@ -63,7 +63,7 @@ void Bedgehog::InitSprites()
 
 }
 
-void Bedgehog::Update(int t)
+void Bedgehog::Update(float t)
 {
 	isCollision = false;
