+++ Metroid/Bedgehog.h
@@ -34,7 +34,7 @@ public:
 	void InitSprites();
 	void ResponseGround(GameObject *target, const float &DeltaTime, const float &CollisionTimeScale);
 	//============== OVERRIDE VIRTUAL METHOD ===================
-	virtual void Update(int t);
+	virtual void Update(float t);
 	virtual void Render();
 	//============== END OVERRIDE VIRTUAL METHOD ===============
 };
